+++ R/ci_profile_boot.R
@@ -125,11 +125,11 @@
 
 
 #' @keywords internal
-.ci_boot_merMod <- function(x, ci, ...) {
+.ci_boot_merMod <- function(x, ci, iterations = 500, effects = "fixed", ...) {
   insight::check_if_installed("lme4")
 
   # Compute
-  out <- as.data.frame(lme4::confint.merMod(x, level = ci, method = "boot", ...))
+  out <- as.data.frame(lme4::confint.merMod(x, level = ci, method = "boot", nsim = iterations, ...))
   rownames(out) <- gsub("`", "", rownames(out), fixed = TRUE)
   out <- out[rownames(out) %in% insight::find_parameters(x, effects = "fixed")$conditional, ]
   names(out) <- c("CI_low", "CI_high")